Refactoring still in process
