+++ src/app/Language.ts
@@ -1,0 +1,1 @@
+export type Language = 'es' | 'en';
