+++ src/assets/sass/styles.css
@@ -1,0 +1,70 @@
+.container {
+  font-family: Raleway, sans-serif;
+}
+
+#calendar-header {
+  text-align: center;
+  background: #0e717f;
+  color: #fff;
+  width: 100%;
+  font-size: 18px;
+  line-height: 20px;
+  font-weight: 700;
+  text-transform: uppercase;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+}
+
+#calendar-header #days .cell {
+  background: #0e717f;
+  color: #fff;
+  -ms-flex-wrap: wrap;
+      flex-wrap: wrap;
+}
+
+#calendar-header :is(button, span):nth-child(-n+3) {
+  -webkit-box-flex: 1;
+      -ms-flex: 1 3 35%;
+          flex: 1 3 35%;
+}
+
+#days-container {
+  border: #CCCCCC 1px solid;
+}
+
+#days-container .cell_empty {
+  width: calc(100% / 7);
+  display: inline-block;
+  text-align: center;
+}
+
+#days-container .cell {
+  border: #CCCCCC 1px solid;
+}
+
+#days-container .cell:active:not(.cell_disabled) {
+  color: white;
+  background-color: #0e717f;
+}
+
+#days-container .cell:hover:not(.cell_disabled) {
+  background-color: #eee;
+}
+
+#days-container .cell_disabled {
+  background: #ddd;
+}
+
+.cell {
+  background: #fff;
+  color: #000;
+  -webkit-box-sizing: border-box;
+          box-sizing: border-box;
+  display: inline-block;
+  padding: 10px 0;
+  text-align: center;
+  width: calc(100% / 7);
+  cursor: pointer;
+}
+/*# sourceMappingURL=styles.css.map */
